Changed its package name for moved classes
