+++ src/main/java/com/mitsuki/jmatrix/exception/IllegalMatrixSizeException.java
@@ -2,7 +2,7 @@
 /* --  IllegalMatrixSizeException  -- */
 // :: ---------------------------- :: //
 
-package com.mitsuki.jmatrix.core;
+package com.mitsuki.jmatrix.exception;
 
 public class IllegalMatrixSizeException extends JMBaseException
 {
