+++ src/main/java/com/mitsuki/jmatrix/exception/InvalidIndexException.java
@@ -2,7 +2,7 @@
 /* --  InvalidIndexException  -- */
 // :: ----------------------- :: //
 
-package com.mitsuki.jmatrix.core;
+package com.mitsuki.jmatrix.exception;
 
 public class InvalidIndexException extends JMBaseException
 {
